requirements.txt
# API Framework
fastapi
uvicorn[standard]

# Data & Finance Libraries
yfinance
pandas
rapidfuzz
tradingview-ta
mplfinance
TA-Lib

# AI & Utilities
google-generativeai
python-dotenv

# News & Web
feedparser
requests
gnews
finnhub-python